+++ commonmark/src/test/java/org/commonmark/test/ParserTest.java
@@ -11,10 +11,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.nio.charset.Charset;
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
+import java.util.*;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
@@ -73,6 +70,12 @@ public class ParserTest {
         parser = Parser.builder().enabledBlockTypes(noCoreTypes).build();
         document = parser.parse(given);
         assertThat(document.getFirstChild(), not(instanceOf(Heading.class)));
+    }
+
+    @Test(expected = IllegalArgumentException.class)
+    public void enabledBlockTypesThrowsWhenGivenUnknownClass() {
+        // BulletList can't be enabled separately at the moment, only all ListBlock types
+        Parser.builder().enabledBlockTypes(new HashSet<>(Arrays.asList(Heading.class, BulletList.class))).build();
     }
 
     @Test